[MINVOKER-67] Create interpolated settings file in cloneProjectsTo directory and keep it after execution
git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk/maven-invoker-plugin@698174 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/maven/plugin/invoker/InvokerMojo.java
@@ -922,8 +922,15 @@ public class InvokerMojo
         File interpolatedSettingsFile = null;
         if ( settingsFile != null )
         {
-            interpolatedSettingsFile =
-                new File( settingsFile.getParentFile(), "interpolated-" + settingsFile.getName() );
+            if ( cloneProjectsTo != null )
+            {
+                interpolatedSettingsFile = new File( cloneProjectsTo, "interpolated-" + settingsFile.getName() );
+            }
+            else
+            {
+                interpolatedSettingsFile =
+                    new File( settingsFile.getParentFile(), "interpolated-" + settingsFile.getName() );
+            }
             buildInterpolatedFile( settingsFile, interpolatedSettingsFile );
         }
 
@@ -944,7 +951,7 @@ public class InvokerMojo
         }
         finally
         {
-            if ( interpolatedSettingsFile != null )
+            if ( interpolatedSettingsFile != null && cloneProjectsTo == null )
             {
                 interpolatedSettingsFile.delete();
             }
